Move most of the rendering logic from the public route into their own controller files
